chore(server): moved tests into server
- move test directory into server folder

[Fixes #161464452]
