+++ server/test/products.test.js
@@ -1,6 +1,6 @@
 import chai from "chai";
 import chaiHttp from "chai-http";
-import server from "../server/";
+import server from "../index";
 
 const should = chai.should();
